+++ lib/groupdate/adapters/sqlite_adapter.rb
@@ -2,9 +2,6 @@ module Groupdate
   module Adapters
     class SQLiteAdapter < BaseAdapter
       def group_clause
-        raise Groupdate::Error, "Time zones not supported for SQLite" unless @time_zone.utc_offset.zero?
-        raise Groupdate::Error, "day_start not supported for SQLite" unless day_start.zero?
-
         query =
           if period == :week
             ["strftime('%Y-%m-%d', #{column}, '-6 days', ?)", "weekday #{(week_start + 1) % 7}"]
@@ -36,7 +33,7 @@ module Groupdate
               when :month
                 "%Y-%m-01"
               when :quarter
-                raise Groupdate::Error, "Quarter not supported for SQLite"
+                nil
               else # year
                 "%Y-01-01"
               end
@@ -44,7 +41,43 @@ module Groupdate
             ["strftime(?, #{column})", format]
           end
 
+        if !@time_zone.utc_offset.zero? || !day_start.zero? || period == :quarter
+          setup_function
+          day_start = self.day_start != 0 ? self.day_start / 3600.0 : nil
+          week_start = period == :week ? Groupdate::Magic::DAYS[self.week_start].to_s : nil
+          query = ["groupdate(?, #{column}, ?, ?, ?)", period, @time_zone.tzinfo.name, day_start, week_start]
+        end
+
         @relation.send(:sanitize_sql_array, query)
+      end
+
+      private
+
+      def setup_function
+        @relation.with_connection do |connection|
+          raw_connection = connection.raw_connection
+          return if raw_connection.instance_variable_defined?(:@groupdate_function)
+
+          utc = ActiveSupport::TimeZone["UTC"]
+          # TODO improve performance
+          raw_connection.create_function("groupdate", 4) do |func, period, value, time_zone, day_start, week_start|
+            if value.nil?
+              func.result = nil
+            else
+              options = {time_zone: time_zone}
+              options[:day_start] = day_start if day_start
+              options[:week_start] = week_start if week_start
+              result = [value].group_by_period(period, **options) { |v| utc.parse(v) }.keys[0]
+              if result.is_a?(Time)
+                result = result.in_time_zone(utc).strftime("%Y-%m-%d %H:%M:%S")
+              elsif result.is_a?(Date)
+                result = result.strftime("%Y-%m-%d")
+              end
+              func.result = result
+            end
+          end
+          raw_connection.instance_variable_set(:@groupdate_function, true)
+        end
       end
     end
   end