+++ test/test_helper.rb
@@ -110,10 +110,6 @@ class Minitest::Test
   def call_method(method, field, options)
     if enumerable?
       @users.group_by_period(method, **options) { |u| u.send(field) }.to_h { |k, v| [k, v.size] }
-    elsif sqlite? && (method == :quarter || (options[:time_zone] && options[:time_zone] != "bad") || options[:day_start] || (Time.zone && options[:time_zone] != false))
-      error = assert_raises(Groupdate::Error) { User.group_by_period(method, field, **options).count }
-      assert_includes error.message, "not supported for SQLite"
-      skip
     else
       User.group_by_period(method, field, **options).count
     end